:scissors: Separate shared test project for mocking and more
Fixes #447
