+++ fuel-android/build.gradle.kts
@@ -3,5 +3,5 @@ dependencies {
     implementation(Kotlin.stdlib)
     testImplementation(RoboElectric.dependency)
     testImplementation(JUnit.dependency)
-    testCompile(MockServer.dependency)
+    testCompile(project(":fuel-test"))
 }
