+++ fuel-forge/build.gradle.kts
@@ -3,5 +3,5 @@ plugins { java }
 dependencies {
     compile(project(":fuel"))
     compile(Forge.dependency)
-    testCompile(MockServer.dependency)
+    testCompile(project(":fuel-test"))
 }
