+++ fuel-jackson/build.gradle.kts
@@ -3,5 +3,5 @@ plugins { java }
 dependencies {
     compile(project(":fuel"))
     compile(Jackson.dependency)
-    testCompile(MockServer.dependency)
+    testCompile(project(":fuel-test"))
 }
