+++ fuel-livedata/build.gradle.kts
@@ -4,5 +4,5 @@ dependencies {
     implementation(Android.Arch.extensions)
     testImplementation(RoboElectric.dependency)
     testImplementation(JUnit.dependency)
-    testCompile(MockServer.dependency)
+    testCompile(project(":fuel-test"))
 }
