+++ fuel-moshi/src/test/kotlin/com/github/kittinunf/fuel/FuelMoshiTest.kt
@@ -6,7 +6,9 @@ import com.github.kittinunf.fuel.core.Request
 import com.github.kittinunf.fuel.core.Response
 import com.github.kittinunf.fuel.moshi.moshiDeserializerOf
 import com.github.kittinunf.fuel.moshi.responseObject
+import com.github.kittinunf.fuel.test.MockHelper
 import com.github.kittinunf.result.Result
+import com.squareup.moshi.Moshi
 import org.hamcrest.CoreMatchers.instanceOf
 import org.hamcrest.CoreMatchers.notNullValue
 import org.junit.After
@@ -14,7 +16,6 @@ import org.junit.Assert.assertNotEquals
 import org.junit.Assert.assertThat
 import org.junit.Before
 import org.junit.Test
-import com.squareup.moshi.Moshi
 import java.net.HttpURLConnection
 
 class FuelMoshiTest {
