+++ fuel-reactor/build.gradle.kts
@@ -6,6 +6,6 @@ dependencies {
     compile(project(":fuel"))
     compile(Reactor.core)
     testCompile(Reactor.test)
-    testCompile(MockServer.dependency)
+    testCompile(project(":fuel-test"))
     testCompile(project(":fuel-jackson"))
 }
