+++ fuel-test/build.gradle.kts
@@ -1,0 +1,7 @@
+plugins { java }
+
+dependencies {
+    compile(project(":fuel"))
+    compile(MockServer.dependency)
+    compile(Json.dependency)
+}
