+++ fuel/build.gradle.kts
@@ -3,5 +3,5 @@ plugins { java }
 dependencies {
     compile(Result.dependency)
     testCompile(Json.dependency)
-    testCompile(MockServer.dependency)
+    testCompile(project(":fuel-test"))
 }
