+++ fuel/src/test/kotlin/com/github/kittinunf/fuel/RequestHandlerTest.kt
@@ -6,6 +6,7 @@ import com.github.kittinunf.fuel.core.Handler
 import com.github.kittinunf.fuel.core.Method
 import com.github.kittinunf.fuel.core.Request
 import com.github.kittinunf.fuel.core.Response
+import com.github.kittinunf.fuel.test.MockHttpTestCase
 import junit.framework.TestCase.assertTrue
 import junit.framework.TestCase.fail
 import org.hamcrest.CoreMatchers.containsString
