+++ fuel/src/test/kotlin/com/github/kittinunf/fuel/issues/ContentTypeHeaderIssue473.kt
@@ -1,9 +1,9 @@
 package com.github.kittinunf.fuel.issues
 
-import com.github.kittinunf.fuel.MockHttpTestCase
-import com.github.kittinunf.fuel.MockReflected
 import com.github.kittinunf.fuel.core.Headers
 import com.github.kittinunf.fuel.core.Method
+import com.github.kittinunf.fuel.test.MockHttpTestCase
+import com.github.kittinunf.fuel.test.MockReflected
 import org.hamcrest.CoreMatchers
 import org.hamcrest.CoreMatchers.equalTo
 import org.hamcrest.MatcherAssert.assertThat
