+++ settings.gradle.kts
@@ -33,5 +33,6 @@ include(":fuel-android")
 include(":fuel-jackson")
 include(":fuel-kotlinx-serialization")
 include(":fuel-moshi")
+include(":fuel-test")
 include(":sample")
 include(":sample-java")
